+++ myems-api/app.py
@@ -542,6 +542,12 @@ api.add_route('/users/changepassword',
               user.ChangePassword())
 api.add_route('/users/unlock/{id_}',
               user.Unlock())
+api.add_route('/users/forgotpassword',
+              user.ForgotPassword())
+api.add_route('/users/emailmessages',
+              user.EmailMessageCollection())
+api.add_route('/users/emailmessages/{id_}',
+              user.EmailMessageItem())
 
 api.add_route('/virtualmeters',
               virtualmeter.VirtualMeterCollection())